pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "fitz>=0.0.1.dev2",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.33.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pymupdf>=1.26.4",
    "reportlab>=4.4.3",
    "sift-stack-py>=0.8.5",
    "werkzeug>=3.1.3",
]
